errorCode优化（3）

+++ backend/yudao-service/yudao-service-biz/src/main/java/cn/iocoder/yudao/service/model/infra/codegen/InfraInterfaceModule.java
@@ -21,7 +21,7 @@ public interface InfraInterfaceModule extends BaseEntity {
 
     Integer type();
 
-    Integer sort();
+    Long sort();
 
     @Nullable
     @LogicalDeleted("now")
